# backend/api/openapi.yaml
openapi: 3.1.0
...
info:
  title: kubeport API
  version: "0.1.0"
  license:
    name: MIT
    identifier: MIT
  description: |
    kubeport turns Kubernetes resources into a self-service catalog: admins
    publish versioned templates, everyone else deploys them through generated
    forms.

    **This API is a UX layer, not an authorization boundary.** Every write that
    touches a cluster is performed with the caller's own OIDC id_token, forwarded
    unchanged to that cluster's apiserver. Kubernetes RBAC decides what actually
    happens, so a 201 from kubeport only means kubeport accepted the request —
    the cluster may still refuse it.

    ## Two ways in

    The Go API described here listens on `:8080` **inside the cluster only**
    (`ClusterIP`; the Ingress routes all external traffic to the frontend). It
    takes an OIDC id_token as a bearer token.

    Browsers instead talk to the Next.js BFF at `https://<host>/api/v1/*`, which
    authenticates from an httpOnly session cookie and injects the bearer token
    server-side. It **ignores any `Authorization` header you send**. Paths and
    bodies are otherwise identical to this spec.

    So a script port-forwards to the backend Service and sends a bearer
    id_token. Driving the browser session is **not** a client path: the BFF
    ignores `Authorization`, and the session token never leaves the server (it
    is held encrypted in the `sessions` table, not in the cookie).

...
    which, and issue #34 tracks the gap.

...

...
servers:
...
security:
  - bearerAuth: []
...
paths:
...
  /v1/me:
    get:
      tags: [identity]
      operationId: getMe
      summary: The caller as kubeport sees them
      description: |
        Use this to discover whether the caller is an admin: `groups` contains
        `kubeport-admin`. That group comes from the IdP's `groups` claim, or —
        when the IdP does not emit one, as with Google — from the
        `KBP_DEV_ADMIN_EMAILS` allowlist.

        The allowlist is a bootstrap escape hatch, not a role system: the
        backend logs a warning at startup when it is set
        (`cmd/server/main.go`), and enabling demo mode makes the chart append
        the public demo admin's address to it
        (`templates/_helpers.tpl`, `kubeport.devAdminEmails`). Keep it to the
        smallest possible list, and empty it once the IdP emits groups.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [subject, email, groups]
                properties:
                  subject: { type: string }
                  email: { type: string, format: email }
                  groups:
                    type: array
                    items: { type: string }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "500": { $ref: "#/components/responses/InternalError" }
  /v1/clusters:
    get:
      tags: [clusters]
      operationId: listClusters
      summary: List clusters available as deploy targets
      description: |
        Returns only what is needed to target a deploy. `api_url`, `ca_bundle`
        and `oidc_issuer_url` are deliberately **not** included for any caller,
        admin included — see issue #52. Register a cluster to see them once, in
        the POST response.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [clusters]
                properties:
                  clusters:
                    type: array
                    items: { $ref: "#/components/schemas/ClusterSummary" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "500": { $ref: "#/components/responses/InternalError" }
    post:
      tags: [clusters]
      operationId: createCluster
      summary: Register a cluster as a deploy target
      description: |
        Admin only, and refused for demo-domain callers. There is no
        cluster-registration screen in the admin UI, so this endpoint is the
        supported path — see `deploy/helm/kubeport/README.md`, "After install".
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateClusterRequest" }
      responses:
        "201":
          description: |
            Created. This is the only response that carries the connection
            fields back, so keep it if you want to verify what was stored.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Cluster" }
        "400": { $ref: "#/components/responses/ValidationError" }
        "413": { $ref: "#/components/responses/PayloadTooLarge" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409":
          description: A cluster with that name already exists.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Problem" }
        "500": { $ref: "#/components/responses/InternalError" }
  /v1/clusters/{name}/openapi:
    get:
      tags: [clusters]
      operationId: getClusterOpenAPIIndex
      summary: Group/version index from the target cluster
      description: |
        Proxies the cluster's own OpenAPI discovery, used by the template editor
        to offer resource kinds. Served from an in-process LRU cache.

        **Rate limited**, sharing one per-caller budget with the sibling `{gv}`
        route and `POST /v1/selfsubjectaccessreview`: a cache miss pulls up to
        10 MiB from the apiserver, and a caller can force misses at will by
        varying `gv` (issue #97).

        The cluster's own status codes are **not** passed through, apart from
        404 — see the `{gv}` route for the mapping, which is shared.

        **What a caller reads depends on who they are** (#124, #283), on this
        route and on `{gv}` alike:

        - a kubeport admin on a non-demo account, or an editor of at least one
          team — the callers who can author templates — reads the cluster's
          whole surface, installed CRDs included;
        - a demo-domain account reads only what the demo RBAC grants: `v1`,
          `apps/v1`, `batch/v1`, `authorization.k8s.io/v1`;
        - any other signed-in caller reads only the built-in workload
          group/versions: `v1`, `apps/v1`, `batch/v1`, `networking.k8s.io/v1`.

        The index lists only what the caller may read. A group/version they may
        not read answers exactly like one the cluster does not have
        (`404 k8s-error` with the fixed detail), and the cluster is not asked.
      parameters:
        - $ref: "#/components/parameters/ClusterName"
      responses:
        "200":
          description: Passed through from the cluster.
          content:
            application/json:
              schema: { type: object, additionalProperties: true }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "404":
          description: |
            Either kubeport has no such cluster (`not-found`) or the cluster's
            OpenAPI root answered 404 (`k8s-error`). Branch on `title` — the
            two routes share one handler, so both 404s occur here too.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Problem" }
        "429": { $ref: "#/components/responses/RateLimited" }
        "502": { $ref: "#/components/responses/ClusterOpenAPIUpstream" }
        "500": { $ref: "#/components/responses/InternalError" }
...
  /v1/clusters/{name}/openapi/refresh:
    post:
      tags: [clusters]
      operationId: refreshClusterOpenAPI
      summary: Drop the cached schema for this cluster
      description: |
        Admin only, and refused for demo-domain callers (issue #97) — the same
        gate the other cluster write carries. Clearing the cache makes the next
        read pull the full schema from the apiserver, up to 10 MiB, so an
        ungated version let any authenticated caller amplify load onto the
        control plane.
      parameters:
        - $ref: "#/components/parameters/ClusterName"
      responses:
        "204": { description: "Cache cleared. No body." }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "500": { $ref: "#/components/responses/InternalError" }
  /v1/selfsubjectaccessreview:
    post:
      tags: [clusters]
      operationId: checkSelfSubjectAccess
      summary: Ask the cluster whether the caller may do something
      description: |
        Forwards a `SelfSubjectAccessReview` using the caller's own token. This
        is how the deploy form greys out actions before submitting — prefer it
        over attempting the write and handling the failure.

        SSAR grants nothing: it reports the caller's own access, and the
        cluster's authorizer decides. What is bounded is the **shape and the
        rate** of the question, because each one becomes a real call to the
        apiserver (issue #73):

        - `verb` must be one kubeport itself performs — see the enum on
          `SSARRequest`. `deletecollection` is in it because that is what
          release deletion actually needs, and k8s authorizes it under its own
          verb rather than under `delete`.
        - `(group, resource)` must be one of the MVP resources kubeport
          applies. Asking about anything else is a 400 whose `detail` **lists
          the allowed set**, so the rule is learnable without guessing.
        - Over budget is a 429 with `Retry-After`. The budget is shared with the
          two cluster-openapi read routes.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SSARRequest" }
      responses:
        "200":
          description: The cluster's verdict.
          content:
            application/json:
              schema:
                type: object
                required: [allowed, denied, reason]
                properties:
                  allowed: { type: boolean }
                  denied: { type: boolean }
                  reason:
                    type: string
                    description: |
                      **Do not branch on this.** It is populated only for
                      kubeport admins on non-demo accounts, and is the empty
                      string for everyone else — including demo accounts, which
                      hold `kubeport-admin` for UX reasons (issue #102).

                      When present it is the cluster authorizer's own sentence,
                      which names the ClusterRoleBinding, ClusterRole and User
                      involved. SSAR is open to any authenticated caller and
                      reports only that caller's own access, so nothing is
                      escalated by it; what leaked was the cluster's binding
                      layout, one namespace at a time.

                      Act on `allowed` / `denied` instead.
        "400":
          description: |
            Body malformed, a required field missing, or `verb` /
            `(group, resource)` outside the allowed sets — `detail` names the
            set in the latter two cases.
          headers:
            X-Request-Id: { $ref: "#/components/headers/RequestId" }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Problem" }
        "413": { $ref: "#/components/responses/PayloadTooLarge" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "404":
          description: No cluster is registered under that name.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Problem" }
        "429": { $ref: "#/components/responses/RateLimited" }
        "500": { $ref: "#/components/responses/InternalError" }
        "502": { $ref: "#/components/responses/UpstreamError" }
  /v1/templates:
    get:
      tags: [templates]
      operationId: listTemplates
      summary: List templates
      description: |
        Each row carries `current_ui_spec`, so one call is enough to render a
        deploy form — no follow-up per template.

        **Not paginated** (issue #58): the whole table is returned.

        Filters narrow what the caller may already see (#74). A query parameter
        not listed here is refused with 400 rather than ignored.
      parameters:
        - name: search
          in: query
          required: false
          description: Case-insensitive substring of `name` or `display_name`.
          schema: { type: string }
        - name: tag
          in: query
          required: false
          description: |
            Keeps templates carrying this tag. Repeat it to require every one
            (`?tag=web&tag=demo`).
          style: form
          explode: true
          schema: { type: array, items: { type: string } }
        - name: status
          in: query
          required: false
          description: |
            The current version's status — `published` or `deprecated` — or
            `draft` for a template never published. Anything else matches
            nothing.
          schema: { type: string, enum: [published, deprecated, draft] }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [templates]
                properties:
                  templates:
                    type: array
                    items: { $ref: "#/components/schemas/TemplateListItem" }
        "400":
          description: A query parameter this endpoint does not take (#74).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Problem" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "500": { $ref: "#/components/responses/InternalError" }
    post:
      tags: [templates]
      operationId: createTemplate
      summary: Create a template and its first version
      description: |
        `authoring_mode` decides which body fields are read: `yaml` uses
        `resources_yaml` + `ui_spec_yaml`, `ui` uses `ui_state`. The first
        version is created as a draft.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateTemplateRequest" }
      responses:
        "201":
          description: Created.
          content:
            application/json:
              schema:
                type: object
                properties:
                  template: { $ref: "#/components/schemas/Template" }
                  version: { $ref: "#/components/schemas/TemplateVersion" }
                  resources_yaml: { type: string }
                  ui_spec_yaml: { type: string }
        "400": { $ref: "#/components/responses/ValidationError" }
        "413": { $ref: "#/components/responses/PayloadTooLarge" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409":
          description: A template with that name already exists.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Problem" }
        "429": { $ref: "#/components/responses/RateLimited" }
        "500": { $ref: "#/components/responses/InternalError" }
  /v1/templates/preview:
    post:
      tags: [templates]
      operationId: previewTemplate
      summary: Convert UI-mode state to YAML without saving
      description: |
        Pure transform of the posted ui_state — reads no template, cluster or
        database — so it is open to any authenticated caller.

        400 when the ui_state would build a document far larger than the body
        that described it. Three limits, all per request: at most 8192 objects
        created, nothing more than 128 levels deep counting a field's path and
        its value together, and at most 1MiB of resources.yaml emitted. Real
        manifest paths run about nine deep and no template approaches any of
        them; a request that trips one is malformed rather than large
        (issue #135).

        Rate limited on a budget shared with the other routes that serialize a
        ui_state, separate from the control-plane routes, and sized for the
        editor's own typing rather than for apiserver fan-out.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [ui_state]
              properties:
                ui_state: { type: object, additionalProperties: true }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  resources_yaml: { type: string }
                  ui_spec_yaml: { type: string }
        "400": { $ref: "#/components/responses/ValidationError" }
        "413": { $ref: "#/components/responses/PayloadTooLarge" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "429": { $ref: "#/components/responses/RateLimited" }
        "500": { $ref: "#/components/responses/InternalError" }
...
  /v1/templates/{name}/versions/{v}/publish:
    post:
      tags: [templates]
      operationId: publishTemplateVersion
      summary: Publish a draft
      description: |
        Makes this the template's current version. Existing releases stay pinned
        to the version they were deployed from.

        Demo accounts get 403 `demo-restricted` for any template unless the
        install set demo.allowTemplateCreate (#294) — the same gate as creating a
        template. Drafts stay open to them.
      parameters:
        - $ref: "#/components/parameters/TemplateName"
        - $ref: "#/components/parameters/VersionNumber"
      responses:
        "200":
          description: Published.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TemplateVersion" }
        "400": { $ref: "#/components/responses/ValidationError" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409":
          description: |
            The version is not in the state this transition requires; `detail`
            names the state it is actually in.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Problem" }
        "500": { $ref: "#/components/responses/InternalError" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        An OIDC **id_token** (not an access token) from an issuer this install
        trusts. The backend checks signature, issuer and `aud`, then forwards
        the same token to the target cluster.

        How to obtain one non-interactively is not solved for production IdPs —
        see `docs/machine-clients.md` and issue #34.
  parameters:
    ClusterName:
      name: name
      in: path
      required: true
      description: The cluster's registered `name`, not its UUID.
      schema: { type: string }
...